Version 1.0.2 - Rework isFullScreen
We get screen physical height from native and compare it with screen height from dart. If difference < 25px, then it is fullscreen.

Note:
We name it fullscreen mode but it is not really full screen. Here we check if user disabled buttons from system navbar, which is not the same of being full screen (edge-edge)